requirements.txt
fastapi
uvicorn
tensorflow-cpu
numpy
Pillow
python-multipart